Refactor "DAEXI_fork-feature"
